+++ clang/lib/Frontend/Rewrite/RewriteObjC.cpp
@@ -213,7 +213,7 @@ namespace {
         return;
 
       // Measure the old text.
-      int Size = Rewrite.getRangeSize(SrcRange);
+      int64_t Size = Rewrite.getRangeSize(SrcRange);
       if (Size == -1) {
         Diags.Report(Context->getFullLoc(Old->getBeginLoc()), RewriteFailedDiag)
             << Old->getSourceRange();
@@ -246,7 +246,7 @@ namespace {
       Diags.Report(Context->getFullLoc(Loc), RewriteFailedDiag);
     }
 
-    void ReplaceText(SourceLocation Start, unsigned OrigLength,
+    void ReplaceText(SourceLocation Start, uint64_t OrigLength,
                      StringRef Str) {
       // If removal succeeded or warning disabled return with no warning.
       if (!Rewrite.ReplaceText(Start, OrigLength, Str) ||
